requirements.txt
requests>=2.28.0
python-dotenv>=1.0.0
aiogram>=3.0.0
